+++ HISTORY.md
@@ -8,3 +8,9 @@
 ======================
 *   update：使用独立模块@nodejs/spider-common和@nodejs/spider-logger； 
 *   update：调整package.json依赖模块；
+
+1.1.0 / 2019-03-15
+======================
+*   update：@nodejs/spider-common升级到0.3.2版本；
+*   feature：添加错误处理模块@nodejs/spider-faith；
+*   feature：spider添加jobId用于记录其错误任务；